Move markers' views into separate views sub-package
